Exercise 8 & 9 Completed
Exercise 8 is a rock paper scissors game which I made the computer randomly pick either rock, paper. or scissors. Exercise 9 is a game where the computer randomly picks a number from 1-9 and the player has to try to guess the number while the computer tells the user if they are too high or too low.