Exercise 16
This exercise is more like a project. I am trying to make a password generator that allows the user to tell the program what kinds of characters it would like to include and also say how long they want the password to be. For now, all I have is all the characters available and have it so the user can pick what kinds of characters they want to be included in the password