Exercise 17
This program is meant to be a web scraper that displays all the names of the article headers in the front page of The New York Times. This is the first time I have done a web scraper and also the first I have used a library and had to read the documentation on it to use it, so this is taking a lot longer than expected, but it is very interesting.