Exercise 18
This program is meant to play a game of cows and bulls with the user. The new concept I learned for this program is having a main() function. I was confused as to how I was able to program without one before, but now I understand and it is interesting to see how it works in python