Exercise 18
I am almost done with exercise 18. I thought I finished but as I was testing it I ran into an issue where if the computer had 2 of the same numbers in it's random number variable, then it would give an extra bull which was confusing because at times it would say 2 cows, 3 bulls. I will fix it next time