Exercise 19 Complete?
The objective of this program/exercise was to print out on 1 page all of the text of an article which was supposedly written on 4 pages. I looked at the article and it was all on 1 page, so I just printed out all the text anyways. I will see if I can work on this program with a different article/web page